Move Constants to constants package
