+++ naming/src/main/java/com/alibaba/nacos/naming/remote/udp/UdpConnector.java
@@ -23,7 +23,7 @@ import com.alibaba.nacos.naming.misc.GlobalExecutor;
 import com.alibaba.nacos.naming.misc.Loggers;
 import com.alibaba.nacos.naming.monitor.MetricsMonitor;
 import com.alibaba.nacos.naming.push.v2.NoRequiredRetryException;
-import com.alibaba.nacos.naming.utils.Constants;
+import com.alibaba.nacos.naming.constants.Constants;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
